% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalVectors.R
\name{computeGlobalVectors}
\alias{computeGlobalVectors}
\title{Create global vectors}
\usage{
computeGlobalVectors(matrix, vectorSize = 300, maxCores = 1)
}
\arguments{
\item{matrix}{A concept co-occurrence matrix as created using \code{\link[=createMatrix]{createMatrix()}}.}

\item{vectorSize}{The number of dimensions of the resulting global vectors.}

\item{maxCores}{The number of parallel cores to use during computation.}
}
\value{
A matrix representing the global vectors. The row names represent the concept IDs.
For your convencience, the concept reference is attached as an attribute.
}
\description{
Create global vectors
}
