% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateBuilder.R
\name{createGloVeCovariateSettings}
\alias{createGloVeCovariateSettings}
\title{Create covariates using GloVe}
\usage{
createGloVeCovariateSettings(
  baseCovariateSettings = createBaseCovariateSettings(),
  conceptVectors,
  analysisIdOffset = 990
)
}
\arguments{
\item{baseCovariateSettings}{The base covariate settings as created using the
\code{createBaseCovariateSettings()} function.}

\item{conceptVectors}{The global concept vectors as created using the
\code{computeGlobalVectors()} function.}

\item{analysisIdOffset}{The first analysis ID to use for the covariates.
Each time window will receive a separate analysis
ID. The last 3 digits of the covariate IDs will be
the analysis ID.}
}
\value{
An object of type \code{covariateSettings}, to be used with prediction models.
}
\description{
Create covariates using GloVe
}
