% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateMatrix.R
\name{createMatrix}
\alias{createMatrix}
\title{Create concept co-occurrence matrix}
\usage{
createMatrix(data, rollUpConcepts = TRUE)
}
\arguments{
\item{data}{An Andromeda object as created using \code{\link[=extractData]{extractData()}}.}

\item{rollUpConcepts}{Should concepts be expanded to include all their ancestors
as well?}
}
\value{
Returns a spare matrix containing the concept co-occurrences. For your
convenience, the concept reference is attached as an attribute.
}
\description{
Create concept co-occurrence matrix
}
