% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GlobalVectors.R
\name{getSimilarConcepts}
\alias{getSimilarConcepts}
\title{Get similar concepts}
\usage{
getSimilarConcepts(conceptId, conceptVectors, n = 25)
}
\arguments{
\item{conceptId}{The concept ID to use as query.}

\item{conceptVectors}{The global concept vectors as created using \code{\link[=computeGlobalVectors]{computeGlobalVectors()}}.}

\item{n}{The number of similar concepts to return.}
}
\value{
Returns a tibble with the concepts most similar to the query concept.
}
\description{
Get similar concepts
}
