% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloVeHd.R
\docType{package}
\name{GloVeHd-package}
\alias{GloVeHd}
\alias{GloVeHd-package}
\title{GloVeHd: Gloval Vectors for HealthCare Data}
\description{
Apply the GloVe algorithm to data in the OMOP Common Data Model.
}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

}
\keyword{internal}
